+++ src/components/maps/Heatmap.tsx
@@ -1,12 +1,13 @@
 import React, { useEffect, useState, useCallback } from 'react'
 import { HeatmapLayer } from '@deck.gl/aggregation-layers'
 import { GeoJsonLayer } from '@deck.gl/layers'
+import { InitialViewStateProps } from '@deck.gl/core/lib/deck'
+import { ColorRange } from '@deck.gl/core/utils/color'
 
 import usaHeatMapData from '../../assets/usa-heatmap.json'
 import BaseMap, { DEFAULT_INITIAL_VIEWSTATE } from './BaseMap'
 import { bboxFromGeoJson, bbox2Viewport } from '../../js/GeoHelpers'
-import { InitialViewStateProps } from '@deck.gl/core/lib/deck'
-import { ColorRange } from '@deck.gl/core/utils/color'
+import { GeojsonFeatureType } from '../../js/types'
 
 const COLOR_RANGE: ColorRange = [
   [1, 152, 189],
@@ -20,7 +21,7 @@ const COLOR_RANGE: ColorRange = [
 const NAV_BAR_OFFSET = 66
 
 interface HeatmapProps {
-  geojson?: { geometry: { type: string } }
+  geojson?: GeojsonFeatureType
   children?: JSX.Element
   getTooltip: (obj: unknown) => void
 }
